requirements.txt
openai>=1.40.0
langchain>=0.3.0,<0.4.0
langchain-openai>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
chromadb>=0.4.22
streamlit>=1.44.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
